Add null check to EnableGirl to prevent errors that would break built limapp. Add ExperienceApp.end call to SceneChange_Timer and rename to GameTimer.
